requirements.txt
flask
flask_mail
flask_login
flask_sqlalchemy
pillow
wheel

